requirements.txt
flask==3.0.0
firebase-admin==6.4.0
pymongo[srv]==4.6.1
python-dotenv==1.0.0
certifi==2024.2.2
flask-socketio==5.3.6
werkzeug==3.0.1
gunicorn==21.2.0
eventlet==0.35.2
dnspython==2.6.1
email-validator==2.1.0.post1
pillow==10.2.0
python-dateutil==2.8.2
pytz==2024.1